+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 Unstable version
 
 - Feat 🥥 Function change: `range` now support single parameter invocation, *compatible with Radash*
+- Fix 🥕 Bug fixed: `deepClone` now processes `DataView.buffer` properly
 - Fix 🥕 Bug fixed: Correct type definition filepath
 
 不稳定版本
 
 - 功能 🥥 变更函数: `range` 现在支持单个参数的调用, *与 Radash 兼容*
+- 修复 🥕 缺陷修复: `deepClone` 现在能正确处理 `DataView.buffer`
 - 修复 🥕 缺陷修复: 修复了类型定义文件路径
 
 ## Version 0.2.0 - 2025-02-16